requirements.txt
flask>=3.1.1
pillow>=11.3.0
python-dotenv>=1.1.1
python-telegram-bot==20.8
requests>=2.32.5
telegram>=0.0.1
bcrypt
psycopg2-binary
gunicorn>=21.2.0
